Update Main.java
Score now displays on the screen; beginning working on a 'console' type area to give player messages, i.e. 'You attacked the Glitch!' 'It missed!' 'You killed the Glitch! +10 points', etc. To do this, the canvas is now the center object of a BorderPane, and the right side is a FlowPane which is moderated by the new Console class. 

Note: even though the score is displaying, the Game Over message is still acting strange when it comes to enemies intersecting with the player.