Added Coins
Two new files: coin class and image. (class will be moved to the appropriate file soon.) Coins are separated from enemy types. Two things that need improved: gold count is not added to the screen yet, and I don't have a way of varying the amounts of gold per coin.